requirements.txt
# 核心依赖
flask>=2.0.0
streamlit>=1.22.0
python-dotenv>=0.19.0

# API调用相关
requests>=2.27.0
openai>=1.0.0

# 其他工具
python-dateutil>=2.8.2 